+++ Time.h
@@ -10,10 +10,10 @@ u8 TimeGetSecFlag(void);
 void TimeSetSecFlag(u8 cmd);
 
 u16 TimeGetLightFlag(void);
-void TimeSetLightFlag(u8 cmd);
+void TimeSetLightFlag(u16 cmd);
 
 u16 TimeGetWaterFlag(void);
-void TimeSetWaterFlag(u8 cmd);
+void TimeSetWaterFlag(u16 cmd);
 
 u16 TimeGetNutritionFlag(void);
 void TimeSetNutritionFlag(u8 cmd);
